Log #1
Made enemy module, with info on where and what.
Fight module changed to show health in colour, and damage in colour.
Created random enemy spawns based on enemy module.
Fixed some issues with enemies spawning on the walls and mountains, and being inaccessible.
Changed how the tiles get their NPCNames, as that was a large part of an error where the enemies spawned but were not drawn.